requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain==0.0.286
langchain_community==0.0.10
openai==1.3.5
tiktoken==0.5.1
pymupdf==1.25.5
python-docx==0.8.11
textract==1.6.5
firebase-admin==6.2.0
# faiss-cpu==1.7.4
chromadb==0.4.13
psycopg2-binary==2.9.10
pydantic-settings
pydantic==2.4.2
python-dotenv==1.0.0
six==1.17.0